+++ processor/src/main/java/react4j/processor/ImmutablePropKeyStrategy.java
@@ -5,5 +5,6 @@ enum ImmutablePropKeyStrategy
   KEYED,
   IS_STRING,
   TO_STRING,
+  ENUM,
   AREZ_IDENTIFIABLE
 }