+++ processor/src/test/resources/expected/com/example/prop/React4j_ImmutablePropTypeEnum.java
@@ -1,6 +1,5 @@
 package com.example.prop;
 
-import arez.Disposable;
 import arez.annotations.ArezComponent;
 import arez.annotations.Feature;
 import arez.annotations.InjectMode;
@@ -18,14 +17,14 @@ import react4j.internal.OnComponentWillUnmount;
 import react4j.internal.OnShouldComponentUpdate;
 
 @ArezComponent(
-    name = "ImmutablePropTypes",
+    name = "ImmutablePropTypeEnum",
     disposeTrackable = Feature.DISABLE,
     allowEmpty = true,
     inject = InjectMode.NONE
 )
 @Generated("react4j.processor.ReactProcessor")
-abstract class React4j_ImmutablePropTypes extends ImmutablePropTypes {
-  React4j_ImmutablePropTypes(@Nonnull final NativeComponent $$react4j$$_nativeComponent) {
+abstract class React4j_ImmutablePropTypeEnum extends ImmutablePropTypeEnum {
+  React4j_ImmutablePropTypeEnum(@Nonnull final NativeComponent $$react4j$$_nativeComponent) {
     bindComponent( $$react4j$$_nativeComponent );
   }
 
@@ -33,49 +32,17 @@ abstract class React4j_ImmutablePropTypes extends ImmutablePropTypes {
   private static ComponentConstructorFunction getConstructorFunction() {
     final ComponentConstructorFunction componentConstructor = ( React.shouldStoreDebugDataAsState() || React.shouldValidatePropValues() ) ? NativeReactComponent::new : LiteNativeReactComponent::new;
     if ( React.enableComponentNames() ) {
-      Js.asPropertyMap( componentConstructor ).set( "displayName", "ImmutablePropTypes" );
+      Js.asPropertyMap( componentConstructor ).set( "displayName", "ImmutablePropTypeEnum" );
     }
     return componentConstructor;
   }
 
   @Override
-  protected ImmutablePropTypes.MyComponent getMyProp() {
+  protected ImmutablePropTypeEnum.Foo getMyProp() {
     if ( React.shouldCheckInvariants() ) {
       return null != props().getAny( Props.myProp ) ? props().getAny( Props.myProp ).cast() : null;
     } else {
       return Js.uncheckedCast( props().getAny( Props.myProp ) );
-    }
-  }
-
-  @Override
-  protected String getMyOtherProp() {
-    if ( React.shouldCheckInvariants() ) {
-      return null != props().getAny( Props.myOtherProp ) ? props().getAny( Props.myOtherProp ).asString() : null;
-    } else {
-      return Js.uncheckedCast( props().getAny( Props.myOtherProp ) );
-    }
-  }
-
-  @Override
-  protected int stillAnotherProp() {
-    return props().getAny( Props.stillAnotherProp ).asInt();
-  }
-
-  @Override
-  protected ImmutablePropTypes.KeyedComponent BobsProp() {
-    if ( React.shouldCheckInvariants() ) {
-      return null != props().getAny( Props.BobsProp ) ? props().getAny( Props.BobsProp ).cast() : null;
-    } else {
-      return Js.uncheckedCast( props().getAny( Props.BobsProp ) );
-    }
-  }
-
-  @Override
-  protected ImmutablePropTypes.Foo getSomeProp() {
-    if ( React.shouldCheckInvariants() ) {
-      return null != props().getAny( Props.someProp ) ? props().getAny( Props.someProp ).cast() : null;
-    } else {
-      return Js.uncheckedCast( props().getAny( Props.someProp ) );
     }
   }
 
@@ -86,18 +53,7 @@ abstract class React4j_ImmutablePropTypes extends ImmutablePropTypes {
   }
 
   private void $$react4j$$_componentWillUnmount() {
-    ((Arez_React4j_ImmutablePropTypes) this).dispose();
-  }
-
-  @Override
-  @Nullable
-  protected ReactNode render() {
-    assert Disposable.isNotDisposed( this );
-    final ImmutablePropTypes.MyComponent $$react4jv$$_getMyProp = getMyProp();
-    if ( Disposable.isDisposed( $$react4jv$$_getMyProp ) ) {
-      return null;
-    }
-    return super.render();
+    ((Arez_React4j_ImmutablePropTypeEnum) this).dispose();
   }
 
   static final class Factory {
@@ -106,23 +62,15 @@ abstract class React4j_ImmutablePropTypes extends ImmutablePropTypes {
 
   static final class Props {
     static final String myProp = React.shouldMinimizePropKeys() ? "a" : "myProp";
-
-    static final String myOtherProp = React.shouldMinimizePropKeys() ? "b" : "myOtherProp";
-
-    static final String stillAnotherProp = React.shouldMinimizePropKeys() ? "c" : "stillAnotherProp";
-
-    static final String BobsProp = React.shouldMinimizePropKeys() ? "d" : "BobsProp";
-
-    static final String someProp = React.shouldMinimizePropKeys() ? "e" : "someProp";
   }
 
   private static final class LiteNativeReactComponent extends NativeComponent implements OnShouldComponentUpdate {
-    private React4j_ImmutablePropTypes $$react4j$$_component;
+    private React4j_ImmutablePropTypeEnum $$react4j$$_component;
 
     @JsConstructor
     LiteNativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
       super( props );
-      $$react4j$$_component = new Arez_React4j_ImmutablePropTypes( this );
+      $$react4j$$_component = new Arez_React4j_ImmutablePropTypeEnum( this );
     }
 
     @Override
@@ -138,12 +86,12 @@ abstract class React4j_ImmutablePropTypes extends ImmutablePropTypes {
   }
 
   private static final class NativeReactComponent extends NativeComponent implements OnShouldComponentUpdate, OnComponentWillUnmount {
-    private React4j_ImmutablePropTypes $$react4j$$_component;
+    private React4j_ImmutablePropTypeEnum $$react4j$$_component;
 
     @JsConstructor
     NativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
       super( props );
-      $$react4j$$_component = new Arez_React4j_ImmutablePropTypes( this );
+      $$react4j$$_component = new Arez_React4j_ImmutablePropTypeEnum( this );
     }
 
     @Override