+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * **\[processor\]** Immutable `@Prop` methods were unable to identify props that implemented `Keyed`
   via implementing an interface that extended `Keyed`. This has been fixed by delegating the decision
   of whether a type is assignable to `Keyed` to the compiler.
+* **\[processor\]** Add support for enum typed immutable props.
 
 ### [v0.116](https://github.com/react4j/react4j/tree/v0.116) (2019-01-31)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.115...v0.116)